c2/cuda/groth16_srs.cuh: refactor SRS load.
- add rigorous boundary checks to on file format;
- save vector sizes next to pointers for future boundary checks;
- coalesce pinned memory allocations;

A note about cudaHostAlloc being performed in the caller thread. The
call effectively suspends execution of the whole program(*), so it makes
no difference if it's moved to the background thread. Having it in
the caller thread on the other hand allows us to react on a failure in a
more direct manner.

(*) Presumably it claims a global lock on TLB tables, so that any thread
    that triggers virtual memory adjustments gets blocked.