Use released bellperson
There is no longer the need to patch `bellperson`. Instead the `cuda-supraseal`
feature is enabled on the Filecoin crates, which makes them use `supraseal-c2`.

This commit also simplifies the dependencies a bit. The `filecoin-proofs-api`
crate isn't really needed.

It also sets `publish = false` on the `c2-test` crate so that it doesn't
accidentally get published.

The `c2-test` can now simply be run by

    cargo test --release -- --nocapture